moved seeding to helpers and removed test migrations
