+++ api/Program.cs
@@ -1,6 +1,6 @@
 using WhaleSpottingBackend.Database;
 using WhaleSpottingBackend.Repositories;
-using WhaleSpottingBackend.Services;
+using WhaleSpottingBackend.Helpers;
 
 var builder = WebApplication.CreateBuilder(args);
